+++ website/src/pages/mod.rs
@@ -4,6 +4,7 @@ mod elements_page;
 mod form_page;
 mod guides_page;
 mod home_page;
+mod layout_page;
 mod not_found_page;
 
 pub use columns_page::ColumnsPage;
@@ -12,4 +13,5 @@ pub use elements_page::ElementsPage;
 pub use form_page::FormPage;
 pub use guides_page::GuidesPage;
 pub use home_page::HomePage;
+pub use layout_page::LayoutPage;
 pub use not_found_page::NotFoundPage;